backend/app/requirements.txt
fastapi
uvicorn
pydantic
httpx
itsdangerous
authlib
PyJWT